+++ backend/domain_events/tests/test_handler.py
@@ -4,25 +4,19 @@ from unittest.mock import patch
 import pytest
 import time_machine
 from conferences.models.speaker_voucher import SpeakerVoucher
-from grants.tests.factories import GrantFactory
 from users.tests.factories import UserFactory
 from schedule.models import ScheduleItem
 from django.utils import timezone
 from pythonit_toolkit.emails.templates import EmailTemplate
 
 from domain_events.handler import (
-    handle_grant_reply_approved_sent,
-    handle_grant_reply_rejected_sent,
-    handle_grant_reply_waiting_list_sent,
     handle_new_cfp_submission,
     handle_new_schedule_invitation_answer,
     handle_schedule_invitation_sent,
     handle_speaker_communication_sent,
     handle_speaker_voucher_email_sent,
     handle_submission_time_slot_changed,
-    handle_grant_voucher_email_sent,
 )
-from grants.models import Grant
 
 
 pytestmark = pytest.mark.django_db
@@ -257,256 +251,6 @@ def test_handle_speaker_voucher_email_sent_cospeaker(settings, speaker_voucher_f
             "is_speaker_voucher": False,
         },
         reply_to=["speakers@placeholder.com"],
-    )
-
-
-@pytest.mark.django_db
-def test_handle_grant_approved_ticket_only_reply_sent(
-    conference_factory, grant_factory, settings
-):
-    settings.FRONTEND_URL = "https://pycon.it"
-
-    conference = conference_factory(
-        start=datetime(2023, 5, 2, tzinfo=timezone.utc),
-        end=datetime(2023, 5, 5, tzinfo=timezone.utc),
-    )
-    user = UserFactory(
-        full_name="Marco Acierno",
-        email="marco@placeholder.it",
-        name="Marco",
-        username="marco",
-    )
-
-    grant = grant_factory(
-        conference=conference,
-        approved_type=Grant.ApprovedType.ticket_only,
-        applicant_reply_deadline=datetime(2023, 2, 1, 23, 59, tzinfo=timezone.utc),
-        total_amount=680,
-        user=user,
-    )
-
-    data = {
-        "grant_id": grant.id,
-        "is_reminder": False,
-    }
-
-    with patch("domain_events.handler.send_email") as email_mock:
-        handle_grant_reply_approved_sent(data)
-
-    email_mock.assert_called_once_with(
-        template=EmailTemplate.GRANT_APPROVED_TICKET_ONLY,
-        to="marco@placeholder.it",
-        subject=f"[{grant.conference.name}] Financial Aid Update",
-        variables={
-            "firstname": "Marco Acierno",
-            "conferenceName": grant.conference.name.localize("en"),
-            "startDate": "2 May",
-            "endDate": "6 May",
-            "deadlineDateTime": "1 February 2023 23:59 UTC",
-            "deadlineDate": "1 February 2023",
-            "replyLink": "https://pycon.it/grants/reply/",
-        },
-        reply_to=["grants@pycon.it"],
-    )
-
-
-@pytest.mark.django_db
-def test_handle_grant_approved_ticket_travel_accommodation_reply_sent(
-    conference_factory, grant_factory, settings
-):
-    settings.FRONTEND_URL = "https://pycon.it"
-
-    conference = conference_factory(
-        start=datetime(2023, 5, 2, tzinfo=timezone.utc),
-        end=datetime(2023, 5, 5, tzinfo=timezone.utc),
-    )
-    user = UserFactory(
-        full_name="Marco Acierno",
-        email="marco@placeholder.it",
-        name="Marco",
-        username="marco",
-    )
-
-    grant = grant_factory(
-        conference=conference,
-        approved_type=Grant.ApprovedType.ticket_travel_accommodation,
-        applicant_reply_deadline=datetime(2023, 2, 1, 23, 59, tzinfo=timezone.utc),
-        travel_amount=680,
-        user=user,
-    )
-    data = {
-        "grant_id": grant.id,
-        "is_reminder": False,
-    }
-
-    with patch("domain_events.handler.send_email") as email_mock:
-        handle_grant_reply_approved_sent(data)
-
-    email_mock.assert_called_once_with(
-        template=EmailTemplate.GRANT_APPROVED_TICKET_TRAVEL_ACCOMMODATION,
-        to="marco@placeholder.it",
-        subject=f"[{grant.conference.name}] Financial Aid Update",
-        variables={
-            "firstname": "Marco Acierno",
-            "conferenceName": grant.conference.name.localize("en"),
-            "startDate": "2 May",
-            "endDate": "6 May",
-            "amount": "680",
-            "deadlineDateTime": "1 February 2023 23:59 UTC",
-            "deadlineDate": "1 February 2023",
-            "replyLink": "https://pycon.it/grants/reply/",
-        },
-        reply_to=["grants@pycon.it"],
-    )
-
-
-@pytest.mark.django_db
-def test_handle_grant_reply_sent_reminder(conference_factory, grant_factory, settings):
-    settings.FRONTEND_URL = "https://pycon.it"
-    conference = conference_factory(
-        start=datetime(2023, 5, 2, tzinfo=timezone.utc),
-        end=datetime(2023, 5, 5, tzinfo=timezone.utc),
-    )
-    user = UserFactory(
-        full_name="Marco Acierno",
-        email="marco@placeholder.it",
-        name="Marco",
-        username="marco",
-    )
-    grant = grant_factory(
-        conference=conference,
-        approved_type=Grant.ApprovedType.ticket_only,
-        applicant_reply_deadline=datetime(2023, 2, 1, 23, 59, tzinfo=timezone.utc),
-        total_amount=680,
-        user=user,
-    )
-    data = {
-        "grant_id": grant.id,
-        "is_reminder": True,
-    }
-
-    with patch("domain_events.handler.send_email") as email_mock:
-        handle_grant_reply_approved_sent(data)
-
-    email_mock.assert_called_once_with(
-        template=EmailTemplate.GRANT_APPROVED_TICKET_ONLY,
-        to="marco@placeholder.it",
-        subject=f"[{grant.conference.name}] Reminder: Financial Aid Update",
-        variables={
-            "firstname": "Marco Acierno",
-            "conferenceName": grant.conference.name.localize("en"),
-            "startDate": "2 May",
-            "endDate": "6 May",
-            "deadlineDateTime": "1 February 2023 23:59 UTC",
-            "deadlineDate": "1 February 2023",
-            "replyLink": "https://pycon.it/grants/reply/",
-        },
-        reply_to=["grants@pycon.it"],
-    )
-
-
-@pytest.mark.django_db
-def test_handle_grant_reply_waiting_list_sent(
-    deadline_factory, conference, grant_factory, settings
-):
-    settings.FRONTEND_URL = "https://pycon.it"
-    user = UserFactory(
-        full_name="Marco Acierno",
-        email="marco@placeholder.it",
-        name="Marco",
-        username="marco",
-    )
-
-    deadline_factory(
-        start=datetime(2023, 3, 1, 23, 59, tzinfo=timezone.utc),
-        conference=conference,
-        type="custom",
-        name={
-            "en": "Update Grants in Waiting List",
-            "it": "Update Grants in Waiting List",
-        },
-    )
-    grant = grant_factory(conference=conference, user=user)
-
-    data = {
-        "grant_id": grant.id,
-    }
-
-    with patch("domain_events.handler.send_email") as email_mock:
-        handle_grant_reply_waiting_list_sent(data)
-
-    email_mock.assert_called_once_with(
-        template=EmailTemplate.GRANT_WAITING_LIST,
-        to="marco@placeholder.it",
-        subject=f"[{grant.conference.name}] Financial Aid Update",
-        variables={
-            "firstname": "Marco Acierno",
-            "conferenceName": grant.conference.name.localize("en"),
-            "replyLink": "https://pycon.it/grants/reply/",
-            "grantsUpdateDeadline": "1 March 2023",
-        },
-        reply_to=["grants@pycon.it"],
-    )
-
-
-@pytest.mark.django_db
-def test_handle_grant_reply_rejected_sent():
-    user = UserFactory(
-        full_name="Marco Acierno",
-        email="marco@placeholder.it",
-        name="Marco",
-        username="marco",
-    )
-    grant = GrantFactory(user=user)
-
-    data = {
-        "grant_id": grant.id,
-    }
-
-    with patch("domain_events.handler.send_email") as email_mock:
-        handle_grant_reply_rejected_sent(data)
-
-    email_mock.assert_called_once_with(
-        template=EmailTemplate.GRANT_REJECTED,
-        to="marco@placeholder.it",
-        subject=f"[{grant.conference.name}] Financial Aid Update",
-        variables={
-            "firstname": "Marco Acierno",
-            "conferenceName": grant.conference.name.localize("en"),
-        },
-        reply_to=["grants@pycon.it"],
-    )
-
-
-def test_handle_grant_voucher_email_sent(settings, grant_factory):
-    user = UserFactory(
-        full_name="Marco Acierno",
-        email="marco@placeholder.it",
-        name="Marco",
-        username="marco",
-    )
-
-    grant = grant_factory(
-        user=user,
-        voucher_code="ABC123",
-    )
-
-    data = {
-        "grant_id": grant.id,
-    }
-
-    with patch("domain_events.handler.send_email") as email_mock:
-        handle_grant_voucher_email_sent(data)
-
-    email_mock.assert_called_once_with(
-        template=EmailTemplate.GRANT_VOUCHER_CODE,
-        to="marco@placeholder.it",
-        subject=f"[{grant.conference.name}] Your Grant Voucher Code",
-        variables={
-            "firstname": "Marco Acierno",
-            "voucherCode": "ABC123",
-        },
-        reply_to=["grants@pycon.it"],
     )
